Add null checks to FileReader, static_cast to FileReader & FileWriter
If the file does not open successfully, or has not finished loading before another operation is attempted on it, then it could attempt to access a null pointer. Null checks are added to prevent this from occurring. This reduces the chance of a crash while attempting to load a file or directory.

static_cast is used instead of C-style casting for improved type safety and explicitness.