Use range-based for loops in Mixer.cpp
myHeaders (used for the WaveOut sound buffer) has a known size, so it's easy and safe to use a range-based loop to remove the need for the index variable.

This is safer in terms of preventing off-by-one or out-of-bounds errors. More specifically, this commit aims to address the case where the audio is briefly missing, usually when starting playback.

In open(MixSource* source, int samplerate), an index variable is still needed, so we now declare it outside of the loop and increment it after each successful iteration.