Fix Tempo Redundant Crash
Previous fix caused a crash when removing a segment by adding a redundant segment only if no other segments of that type were present after that point.

Changing to properly tagging a segment as redundant when overlapping a existing segment fixes this issue.

For example:
If there was only one Label in a file, and you set it to blank using the menu, it would crash.

Deleting it through the tempo boxes would work as intended.