renderWaveform optimization
This groups related operations more clearly, and reduces branching by using a series of if-else statements instead of switch cases for luminance and wave shape. The logic remains the same - no functional differences with this commit. This is most noticeable when loading MP3, but is still a minor improvement with OGG.

Benchmarks:
25MB 128kbps stereo MP3 (45 mins)
this commit: 6.5 seconds
existing code:  about 12 seconds

52MB VBR stereo MP3 (1 hour)
this commit: 8.7 seconds
existing code: 14.6 seconds

120MB 224kbps stereo OGG (1 hour)
this commit: 12.0 seconds
existing code: 12.5 seconds