Make a function for handling load via file drop
The ultimate goal of this commit was to move the handling of loading files via drag-and-drop out of the handleMsg switch case and into its own function.

Once it was in its own function I made the function more robust. I also noticed a frequent crash was due to a race condition between the destruction of WideString and the DragFinish call. Scoping the loop that contains WideString guarantees this race condition can't occur.

Resolve a crash caused by really long path names

https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file#maximum-path-length-limitation

Update System.cpp

Update System.cpp