Don't call glGetError() every frame
based on other internet discussions, this call can be expensive. and if you check the current application in a profiler, this function comes up as being the hottest. removing its use from a release build seems to lower cpu times, so why not?

it is only disabled in the loop here with a define, but is left intact at its root because its use in other less frequent locations can still be useful for user error logging