Implement Select By Column Count/Density
This only select notes that match the density selected on a given row.

This ignores mines that would also be on the same row.

Key Shortcuts were added for 1-10, though the menu will display up to the MAX_SIM_COLUMNS of 32, matching the column count of the current sim file.