Beatline row hover info (#248)
* Add hover info for mouse position.

Keep track of the nearest snapped row when hovering the notefield with the mouse.

Adds a option to display the row/beat/measure/time of the hovered row into the Status Bar.

* Add additional beatline options.

- Use current snap
- Use snap colors
- Highlight mouse position.

This also adds saving/loading of these options along with waveform & note display.

This changes the default shortcut name for Beatline toggle, breaking compatibility with older settings.txt, but makes it consistent with other setting groups.

Additionally snap colors used in minimap drawing have been moved to Common are are reused for the snap colors, with a slight tweaked blue color.

* Update Statusbar.cpp